+++ src/functionalTest/kotlin/com/cmgapps/license/LicensePluginJavaShould.kt
@@ -59,7 +59,7 @@ class LicensePluginJavaShould {
             .withPluginClasspath()
     }
 
-    @ParameterizedTest(name = "{displayName} - Gradle Version={0}")
+    @ParameterizedTest(name = "${ParameterizedTest.DISPLAY_NAME_PLACEHOLDER} - Gradle Version = {0}")
     @ValueSource(strings = ["6.8", "6.9", "7.0", "7.1", "7.2", "7.3", "7.4"])
     fun `apply Licenses plugin to various Gradle versions`(version: String) {
         val result = gradleRunner